+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG.md
 
+## unreleased
+ - added support for the MONEY and SMALLMONEY types in MSSQL.
+
 ## v0.35.2
  - Fix a bug with zero values being displayed with a non-zero height in stacked bar charts.
  - Updated dependencies, including the embedded SQLite database.